fix: rename mantle-wadsley to mantle-testnet
